requirements.txt
# Required Python packages get listed here, one per line.

runpod~=1.7.9

# Stable Diffusion and AI/ML dependencies
diffusers>=0.25.0
torch>=2.1.0
transformers>=4.35.0
accelerate>=0.24.0
safetensors>=0.4.0
pillow>=10.0.0
requests>=2.31.0
